+++ yudao-module-erp/yudao-module-erp-api/src/main/java/cn/iocoder/yudao/module/erp/api/product/dto/ErpCustomRuleDTO.java
@@ -1,5 +1,6 @@
 package cn.iocoder.yudao.module.erp.api.product.dto;
 
+import lombok.Builder;
 import lombok.Data;
 
 /**
@@ -8,6 +9,7 @@ import lombok.Data;
  * 产品数据:海关规则-> 1:n
  */
 @Data
+@Builder
 public class ErpCustomRuleDTO {
     /**
      * 海关规则id  =====>>>> 映射到pd_declaration_statement
