+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/controller/admin/logistic/customrule/vo/ErpCustomRuleRespVO.java
@@ -19,8 +19,8 @@ import java.time.LocalDateTime;
 @ExcelIgnoreUnannotated
 public class ErpCustomRuleRespVO implements VO {
 
-    @Schema(description = "产品编号", requiredMode = Schema.RequiredMode.REQUIRED)
-    @ExcelProperty("产品编号")
+    @Schema(description = "海关规则编号", requiredMode = Schema.RequiredMode.REQUIRED)
+    @ExcelProperty("海关规则编号")
     private Long id;
 
     @Schema(description = "国家编码", requiredMode = Schema.RequiredMode.REQUIRED)
