+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/convert/product/ErpProductConvert.java
@@ -1,0 +1,16 @@
+package cn.iocoder.yudao.module.erp.convert.product;
+
+import cn.iocoder.yudao.module.erp.api.product.dto.ErpProductDTO;
+import cn.iocoder.yudao.module.erp.dal.dataobject.product.ErpProductDO;
+import org.mapstruct.Mapper;
+import org.mapstruct.factory.Mappers;
+
+@Mapper
+public interface ErpProductConvert {
+
+    ErpProductConvert INSTANCE = Mappers.getMapper(ErpProductConvert.class);
+
+
+    ErpProductDTO convert(ErpProductDO erpProductDO);
+
+}
