+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/service/logistic/customrule/ErpCustomRuleService.java
@@ -66,4 +66,12 @@ public interface ErpCustomRuleService {
      * @return ErpCustomRuleDTO dto
      */
     ErpCustomRuleDTO listErpCustomRuleDTOById(Long id);
+    /**
+     * 根绝产品id(确保存在)获取n个产品DTO，获取产品+海关规则。如果海关规则无匹配，则返回null
+     * <p>
+     *
+     * @param productId 产品id
+     * List<ErpCustomRuleDTO> 海关规则+产品 DTOs
+     */
+    List<ErpCustomRuleDTO> listErpCustomRuleDTOsByProductId(Long productId);
 }
