+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/service/product/ErpProductService.java
@@ -1,7 +1,6 @@
 package cn.iocoder.yudao.module.erp.service.product;
 
 import cn.iocoder.yudao.framework.common.pojo.PageResult;
-import cn.iocoder.yudao.module.erp.api.product.dto.ErpCustomRuleDTO;
 import cn.iocoder.yudao.module.erp.controller.admin.product.vo.product.ErpProductPageReqVO;
 import cn.iocoder.yudao.module.erp.controller.admin.product.vo.product.ErpProductRespVO;
 import cn.iocoder.yudao.module.erp.controller.admin.product.vo.product.ErpProductSaveReqVO;
@@ -126,12 +125,4 @@ public interface ErpProductService {
      */
     Long getProductCountByUnitId(Long unitId);
 
-    /**
-     * 根绝产品id获取n个产品DTO，获取产品+海关规则。如果海关规则无匹配，则返回null
-     * <p>
-     *
-     * @param productId 产品id
-     * List<ErpCustomRuleDTO> 海关规则+产品 DTOs
-     */
-     List<ErpCustomRuleDTO> listErpCustomRuleDTOsByProductId(Long productId);
 }
