+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/service/product/ErpProductServiceImpl.java
@@ -9,19 +9,16 @@ import cn.iocoder.yudao.framework.common.pojo.PageResult;
 import cn.iocoder.yudao.framework.common.util.collection.MapUtils;
 import cn.iocoder.yudao.framework.common.util.object.BeanUtils;
 import cn.iocoder.yudao.framework.security.core.util.SecurityFrameworkUtils;
-import cn.iocoder.yudao.module.erp.api.product.dto.ErpCustomRuleDTO;
 import cn.iocoder.yudao.module.erp.api.product.dto.ErpProductDTO;
 import cn.iocoder.yudao.module.erp.controller.admin.product.vo.product.ErpProductPageReqVO;
 import cn.iocoder.yudao.module.erp.controller.admin.product.vo.product.ErpProductRespVO;
 import cn.iocoder.yudao.module.erp.controller.admin.product.vo.product.ErpProductSaveReqVO;
 import cn.iocoder.yudao.module.erp.controller.admin.product.vo.product.json.GuidePriceJson;
-import cn.iocoder.yudao.module.erp.convert.logistic.CustomRuleConvert;
-import cn.iocoder.yudao.module.erp.dal.dataobject.logistic.customrule.ErpCustomRuleDO;
 import cn.iocoder.yudao.module.erp.dal.dataobject.product.ErpProductCategoryDO;
 import cn.iocoder.yudao.module.erp.dal.dataobject.product.ErpProductDO;
 import cn.iocoder.yudao.module.erp.dal.dataobject.product.ErpProductUnitDO;
-import cn.iocoder.yudao.module.erp.dal.mysql.logistic.customrule.ErpCustomRuleMapper;
 import cn.iocoder.yudao.module.erp.dal.mysql.product.ErpProductMapper;
+import cn.iocoder.yudao.module.erp.service.logistic.customrule.ErpCustomRuleService;
 import cn.iocoder.yudao.module.erp.service.product.bo.ErpProductBO;
 import cn.iocoder.yudao.module.system.api.dept.DeptApi;
 import cn.iocoder.yudao.module.system.api.dept.dto.DeptRespDTO;
@@ -67,7 +64,7 @@ public class ErpProductServiceImpl implements ErpProductService {
     @Resource
     DeptApi deptApi;
     @Resource
-    ErpCustomRuleMapper customRuleMapper;
+    ErpCustomRuleService erpCustomRuleService;
     @Resource
     AdminUserApi userApi;
 
@@ -183,7 +180,7 @@ public class ErpProductServiceImpl implements ErpProductService {
         ThrowUtil.ifSqlThrow(productMapper.updateById(updateObj),DB_UPDATE_ERROR);
         //更新产品时->覆盖n个海关规则
         //找到产品id对应的所有产品DTO(含海关信息)
-        Optional.ofNullable(this.listErpCustomRuleDTOsByProductId(id)).ifPresent(
+        Optional.ofNullable(erpCustomRuleService.listErpCustomRuleDTOsByProductId(id)).ifPresent(
                 dtos -> erpCustomRuleChannel.send(MessageBuilder.withPayload(dtos).build())
         );
         //获取创建人id
@@ -437,24 +434,5 @@ public class ErpProductServiceImpl implements ErpProductService {
         Integer level = deptApi.getDeptLevel(deptId);
         //判断登记是否符合要求
         ThrowUtil.ifThrow(!levels.contains(level), DEPT_LEVEL_NOT_MATCH);
-    }
-
-    /**
-     * 根绝产品id获取n个海关规则信息 1:n
-     * @param productId 产品id
-     * List<ErpCustomRuleDTO> 海关规则+产品 DTOs
-     */
-    public List<ErpCustomRuleDTO> listErpCustomRuleDTOsByProductId(Long productId) {
-        validateProductExists(productId);
-        //1.0 获得产品
-        ErpProductDO erpProductDO = productMapper.selectById(productId);
-        //2.0 获得海关规则
-        List<ErpCustomRuleDO> erpCustomRuleDOList = customRuleMapper.selectByProductId(productId);
-        //3.0 封装返回
-        if (CollUtil.isNotEmpty(erpCustomRuleDOList)) {
-            return erpCustomRuleDOList.stream().map(erpCustomRuleDO -> CustomRuleConvert.INSTANCE.convert(erpCustomRuleDO, erpProductDO)).collect(Collectors.toList());
-        }else{
-            return Collections.emptyList();
-        }
     }
 }
